Updating to theoretically increase the speed of the database calls by converting three seperate calls into one.

+++ src/main/java/org/smartjobs/adaptors/data/repository/data/CriteriaAnalysis.java
@@ -19,7 +19,7 @@ public class CriteriaAnalysis {
     private Long id;
     private long analysisId;
     private double score;
-    private int maxScore;
+    private long maxScore;
     private String criteriaRequest;
     private String description;
     private long userCriteriaId;
