+++ src/main/java/org/smartjobs/core/entities/ScoredCriteria.java
@@ -1,5 +1,5 @@
 package org.smartjobs.core.entities;
 
 public record ScoredCriteria(long userCriteriaId, String criteriaRequest, String justification, double score,
-                             int maxScore) {
+                             long maxScore) {
 }
