+++ src/main/java/org/smartjobs/core/entities/UserScoringCriteria.java
@@ -6,6 +6,6 @@ public record UserScoringCriteria(long id,
                                   CriteriaCategory category,
                                   String criteriaDescription,
                                   boolean isBoolean,
-                                  int weighting,
+                                  long weighting,
                                   String scoringGuide) {
 }
